JE 0.75
Keyboard and Mouse have the ability to be disabled now.
Audio source adapted for ResourceBundle change now
Added camera shake script
Removed some pointless packages
Made physics less memory intensive.
Camera viewport now auto-adjusts with the window size
Nuklear can be disabled in window preferences
Added a loading screen sequence if you want to show a screen before transitioning to the next scene. (Future updates will go more into depth)
Resource manager can warmup an array of assets and store them by name for quick access before your scene
Added static get method to texture to retrieve a texture by name without doing the whole ResourceManager.getIfExists thing.
Added texture names to SpriteRenderer

+++ src/main/java/org/JE/JE2/Rendering/Renderers/Renderer.java
@@ -6,16 +6,14 @@ import org.JE.JE2.Annotations.GLThread;
 import org.JE.JE2.Annotations.HideFromInspector;
 import org.JE.JE2.Manager;
 import org.JE.JE2.Objects.GameObject;
-import org.JE.JE2.Objects.Scripts.Base.Script;
-import org.JE.JE2.Objects.Scripts.Common.Transform;
+import org.JE.JE2.Objects.Scripts.Script;
+import org.JE.JE2.Objects.Scripts.Transform;
 import org.JE.JE2.Objects.Lights.Light;
 import org.JE.JE2.Rendering.Camera;
 import org.JE.JE2.Rendering.Shaders.ShaderLayout;
 import org.JE.JE2.Rendering.Shaders.ShaderProgram;
 import org.JE.JE2.Rendering.VertexBuffers.VAO;
 import org.JE.JE2.UI.UIElements.Style.Color;
-import org.joml.Vector2f;
-import org.joml.Vector3f;
 import org.lwjgl.BufferUtils;
 
 import java.util.ArrayList;
@@ -64,7 +62,7 @@ public class Renderer extends Script {
 
     @GLThread
     public void Render(GameObject gameObject, int additionalBufferSize, Camera camera) {
-        glViewport(camera.viewportSize.x, camera.viewportSize.y, camera.viewportSize.z, camera.viewportSize.w);
+        glViewport((int) camera.viewportSize.x, (int) camera.viewportSize.y, (int) camera.viewportSize.z, (int) camera.viewportSize.w);
         PreRender();
         ShaderProgram shader = vao.getShaderProgram();
         if(!shader.use())
